requirements.txt
# Web
Flask
gunicorn
flask-cors

# Utilidades
requests
python-dotenv
python-dateutil
pytz

# IA / Voz
openai==0.28.1
pydub==0.25.1
elevenlabs==1.13.5

# Google Cloud / Firebase
firebase-admin
google-cloud-firestore
google-cloud-texttospeech
google-cloud-speech
google-cloud-storage

# Tabelas (importação de preços)
pandas
openpyxl

# Docs (opcional, para leitura .docx no GCS)
python-docx
